CodeBuild: drop configurable execution knobs, merge tests
AGENT_IMAGE, WORKSPACE, and SOURCE_PATH aren't real AWS choices, so
they're fixed constants now instead of env vars; MAX_LOG_EVENTS and
the AWS_ENDPOINT override are dropped outright. Execution tests move
into test_codebuild.py alongside the rest of the CodeBuild suite.
